feat: Add click event to exercise list items in Workout.vue
The code changes in `Workout.vue` add a click event to the exercise list items. When a user clicks on an exercise, the `openExercise` method is called, which navigates to the `exercise` route with the corresponding exercise ID as a parameter.

This commit message follows the established convention of starting with a verb in the imperative form and providing a clear and concise description of the changes made.